requirement.txt
ChatterBot-1.0.5
chatterbot-corpus-1.2.0
simple-websocket-server
Python-3.7